refactor(#4): Ajustando boards.
